Factor out CodeFixes and Analyzers to separate assemblies
