+++ src/Framework/AspNetCoreAnalyzers/src/Analyzers/DiagnosticDescriptors.cs
