+++ src/Framework/AspNetCoreAnalyzers/test/TestDiagnosticAnalyzer.cs
